src/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.51.0
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-etc==1.3.1
django-extensions==3.1.5
django-filter==22.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-recaptcha==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
flake8==4.0.1
fonttools==4.33.3
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
jinja-to-js==3.2.3
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterthemes==0.20.0
kiwisolver==1.4.2
lesscpy==0.15.0
lorem==0.1.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg==3.0.14
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pyzmq==23.0.0
requests==2.28.0
requests-oauthlib==1.3.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
Unidecode==1.3.4
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
